functions/requirements.txt
# CloudSage Azure Functions — Runtime Dependencies (free-tier profile)
# Only packages needed for the HTTP-triggered Function are included.
# Paid-service packages are commented out.

azure-functions>=1.18.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
# azure-servicebus>=7.11.0     # paid — restore for Service Bus trigger
# azure-storage-blob>=12.19.0  # paid — restore for FAISS blob backing
# azure-mgmt-containerservice>=28.0.0  # paid — AKS management

openai>=1.30.0
faiss-cpu>=1.7.4
# kubernetes>=28.1.0            # paid — AKS automation

scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
tenacity>=8.2.0
requests>=2.31.0
